change to use float in output marker poses

+++ marker_detection.cpp
@@ -10,7 +10,7 @@
 // Give out a list of 4x4 transformation matrices (rotation + translation)
 void detectMarkersAndEstimatePose(
 	const cv::Mat& input_image,
-	std::vector<cv::Mat_<double>>& output_marker_poses) {
+	std::vector<cv::Mat_<float>>& output_marker_poses) {
 	if (!output_marker_poses.empty()) {
 		output_marker_poses.clear();
 	}
